+++ conans/test/functional/toolchains/cmake/test_cmake_toolchain.py
@@ -7,8 +7,8 @@ import textwrap
 import pytest
 
 from conan.tools.cmake.presets import load_cmake_presets
-from conans.model.recipe_ref import RecipeReference
 from conan.tools.microsoft.visual import vcvars_command
+from conans.model.recipe_ref import RecipeReference
 from conans.test.assets.cmake import gen_cmakelists
 from conans.test.assets.genconanfile import GenConanfile
 from conans.test.utils.test_files import temp_folder
@@ -1908,93 +1908,3 @@ def test_cmake_toolchain_cxxflags_multi_config():
     c.run_command(r"build\Debug\example.exe")
     assert 'answer=123' in c.out
     assert "CPLUSPLUS: __cplusplus19" in c.out
-
-
-@pytest.mark.tool("cmake")
-@pytest.mark.skipif(platform.system() != "Darwin", reason="need clang and gcc (clang)")
-def test_toolchain_and_compilers_build_context():
-    """
-    Tests how CMakeToolchain manages the build context profile if the build profile is
-    specifying another compiler path (using conf)
-
-    Issue related: https://github.com/conan-io/conan/issues/15878
-    """
-    host = textwrap.dedent("""
-    [settings]
-    arch=armv8
-    build_type=Release
-    compiler=gcc
-    compiler.cppstd=gnu17
-    compiler.libcxx=libstdc++11
-    compiler.version=11
-    os=Macos
-
-    [conf]
-    tools.build:compiler_executables={"c": "gcc", "cpp": "g++"}
-    """)
-    build = textwrap.dedent("""
-    [settings]
-    os=Macos
-    arch=x86_64
-    compiler=clang
-    compiler.version=12
-    compiler.libcxx=libc++
-    compiler.cppstd=11
-
-    [conf]
-    tools.build:compiler_executables={"c": "clang", "cpp": "clang++"}
-    """)
-    tool = textwrap.dedent("""
-    import os
-    from conan import ConanFile
-    from conan.tools.files import replace_in_file
-    from conan.tools.cmake import CMake
-
-    class toolRecipe(ConanFile):
-        name = "tool"
-        version = "1.0"
-        # Binary configuration
-        settings = "os", "compiler", "build_type", "arch"
-        generators = "CMakeToolchain"
-        exports_sources = "*"
-
-        def build(self):
-            b = CMake(self)
-            b.configure()
-    """)
-    consumer = textwrap.dedent("""
-    import os
-    from conan import ConanFile
-    from conan.tools.files import replace_in_file
-    from conan.tools.cmake import CMake
-
-    class consumerRecipe(ConanFile):
-        name = "consumer"
-        version = "1.0"
-        # Binary configuration
-        settings = "os", "compiler", "build_type", "arch"
-        generators = "CMakeToolchain"
-        exports_sources = "*"
-        tool_requires = "tool/1.0"
-
-        def build(self):
-            b = CMake(self)
-            b.configure()
-    """)
-    client = TestClient()
-    client.save({
-        "host": host,
-        "build": build,
-        "tool/conanfile.py": tool,
-        "tool/CMakeLists.txt": "project(tool)\n",
-        "consumer/conanfile.py": consumer,
-        "consumer/CMakeLists.txt": "project(consumer)\n",
-    })
-    client.run("export tool")
-    client.run("create consumer -pr:h host -pr:b build --build=missing")
-    assert "-- Check for working C compiler: /usr/bin/clang - skipped" in str(client.out)
-    assert "-- Check for working CXX compiler: /usr/bin/clang++ - skipped" in str(client.out)
-    assert "-- Check for working C compiler: /usr/bin/gcc - skipped" in str(client.out)
-    assert "-- Check for working CXX compiler: /usr/bin/g++ - skipped" in str(client.out)
-    assert "consumer/1.0: Created package revision" in str(client.out)
-    assert "tool/1.0: Created package revision" in str(client.out)
